Move CredentialService initialization in VSSolutionManager as we need in it init in restore package for autorestore to work with VSO feeds
