+++ src/NuGet.Clients/NuGet.Tools/NuGetPackage.cs
@@ -158,8 +158,6 @@ namespace NuGetVSExtension
             _dteEvents = _dte.Events.DTEEvents;
             _dteEvents.OnBeginShutdown += OnBeginShutDown;
 
-            SetDefaultCredentialProvider();
-
             if (SolutionManager.NuGetProjectContext == null)
             {
                 SolutionManager.NuGetProjectContext = ProjectContext;
@@ -193,108 +191,6 @@ namespace NuGetVSExtension
                     Resx.Label_NuGetWindowCaption,
                     project.Name));
             }
-        }
-
-        /// <summary>
-        /// Set default credential provider for the HttpClient, which is used by V2 sources.
-        /// Also set up authenticated proxy handling for V3 sources.
-        /// </summary>
-        private void SetDefaultCredentialProvider()
-        {
-            var credentialService = GetCredentialService();
-
-            HttpHandlerResourceV3.CredentialService = credentialService;
-        }
-
-        private NuGet.Configuration.ICredentialService GetCredentialService()
-        {
-            // Initialize the credential providers.
-            var credentialProviders = new List<ICredentialProvider>();
-
-            TryAddCredentialProviders(
-                credentialProviders,
-                Resources.CredentialProviderFailed_VisualStudioAccountProvider,
-                () =>
-                {
-                    var importer = new VsCredentialProviderImporter(
-                        _dte,
-                        VisualStudioAccountProvider.FactoryMethod,
-                        (exception, failureMessage) => LogCredentialProviderError(exception, failureMessage));
-
-                    return importer.GetProviders();
-                });
-
-            TryAddCredentialProviders(
-                credentialProviders,
-                Resources.CredentialProviderFailed_VisualStudioCredentialProvider,
-                () =>
-                {
-                    var webProxy = (IVsWebProxy)GetService(typeof(SVsWebProxy));
-
-                    Debug.Assert(webProxy != null);
-
-                    return new NuGet.Credentials.ICredentialProvider[] {
-                        new VisualStudioCredentialProvider(webProxy)
-                    };
-                });
-
-            if (PreviewFeatureSettings.DefaultCredentialsAfterCredentialProviders)
-            {
-                TryAddCredentialProviders(
-                credentialProviders,
-                Resources.CredentialProviderFailed_DefaultCredentialsCredentialProvider,
-                () =>
-                {
-                    return new NuGet.Credentials.ICredentialProvider[] {
-                        new DefaultCredentialsCredentialProvider()
-                    };
-                });
-            }
-
-            // Initialize the credential service.
-            var credentialService = new CredentialService(credentialProviders, nonInteractive: false);
-
-            return credentialService;
-        }
-
-        private void TryAddCredentialProviders(
-            List<NuGet.Credentials.ICredentialProvider> credentialProviders,
-            string failureMessage,
-            Func<IEnumerable<NuGet.Credentials.ICredentialProvider>> factory)
-        {
-            try
-            {
-                var providers = factory();
-
-                if (providers != null)
-                {
-                    foreach (var credentialProvider in providers)
-                    {
-                        credentialProviders.Add(credentialProvider);
-                    }
-                }
-            }
-            catch (Exception exception)
-            {
-                LogCredentialProviderError(exception, failureMessage);
-            }
-        }
-
-        private void LogCredentialProviderError(Exception exception, string failureMessage)
-        {
-            // Log the user-friendly message to the output console (no stack trace).
-            OutputConsoleLogger.Log(
-                MessageLevel.Error,
-                failureMessage +
-                Environment.NewLine +
-                ExceptionUtilities.DisplayMessage(exception));
-
-            // Write the stack trace to the activity log.
-            ActivityLog.LogWarning(
-                ExceptionHelper.LogEntrySource,
-                failureMessage +
-                Environment.NewLine +
-                exception);
         }
 
         private async Task AddMenuCommandHandlersAsync()
