+++ src/NuGet.Clients/PackageManagement.VisualStudio/ICredentialServiceProvider.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.Composition;
+
+namespace NuGet.PackageManagement.VisualStudio
+{
+    public interface ICredentialServiceProvider
+    {
+        NuGet.Configuration.ICredentialService GetCredentialService();
+    }
+}
